Map markers
Fetches map marker data from the server and displays them with the appropriate colors. Falls back on `localhost` for web debug to sidestep CORS denials.

This requires a patch of `google_maps_flutter_android` to render at a reasonable size on Android, as it was not performing any DPI scaling for non-resolution-aware assets.