Switching to a custom port of Joda Time
In preparation for adding persistence.

We were using timezone, which is nice but lacks detailed DST and local time math.

We tried time_machine, which is another port of (N)oda Time, which offers the functionality we want but lacks some quality of life features like interfaces and operators, and looked a bit too heavy to fork to add the features we wanted.

Instead, let's roll our own Joda Time port backed by timezone and the Dart core DateTime arithmetic.