Map tile caching
Adapts the tile cache to use Hive instead of a sparse array, at a single LOD. This simplifies a lot of the logic, but we might want to explore assembling offline placeholders as well if fetches fail.

Also changes app initialization to block on Hive init since we expect that to be pretty fast and it simplifies the logic.