Tide/current station selection for non-tide-current locations
Enables reading the stand-in tide/current station for locations like launches and destinations. These are kept as IDs so that the structure can stay immutable and so that we don't have to do a linking dance, especially since the app will eventually need the ID => Station map anyway for things like deep links.

Does not yet visually mark the graph as being a proxy station.