[map] Switch marker z-ordering depending on open tab
Tide/current stations are more useful in the Tides tab and destinations/put-ins are more useful in the Details tab. This matters most on mobile where touch target precedence makes one or the other easier to select.

This is a relatively expensive operation so it happens after any animations. This requires a patch in Flutter due to a scheduler bug.

Updates the README with the new Flutter fork, and removes notes about requiring CanvasKit since we're bypassing canvas rendering for web tiles now.