Elevate timeWindow to top-level state and keep details panel alive
These changes are meant to provide a more natural experience when switching tabs. Keeping the timeWindow at the top level also preserves it when a location without tide/current data is selected, where even a keep-alive subtree wouldn't do it.

Also changes GraphTimeWindow.t to a DateTime and enforces that its time zone is consistent with t0.

Fixes #15