[map] Use custom controls for location and zoom
Custom controls allow for consistent custom layout. This is in preparation for map orientation, which will customize the My Location button.

Move the custom layout to the left for left-handed use and since the panel prefers to be to the right. Overflow the LOD controls to the lower right.

Disables the compass since it will now be occluded. This functionality will be added to the location control later.