[map] Disable location button for insecure non-localhost
Chrome, Firefox, and Edge do not surface location requests to the user in insecure contexts. Since we expect to be using https in production, display a tooltip for the remote dev server use case. It might be more ideal to hide the button or catch the error, which may be warranted in the future.