[map] Track north on location button double tap from free bearing
This patches a hole in the idiom of tapping on the button in quick succession to get it to the desired state. Previously, if the user was moving around the map with free bearing, tapping on the location button a second time would effectively do nothing.

It might be natural to do this instead if the animation back to north was already in progress when the button was tapped, but the API does not expose that and it's not worth hacking in.