[map] Custom camera animation controller and higher resolution orientation sampling
Use a custom animation controller to coordinate map camera animations. This is required to allow sensible animation behaior at higher frequency updates.

Also switches to a flutter_sensors implementation based on smooth_compass rather than flutter_compass for now to put more of the postprocessing at the Dart level instead of platform-specific. For now this loses us the axis adaptation when the device is pitched or rolled, and the magnetometer/accelerometer fallback on Android.