[map] Fix tracking regression
e36f05 introduced a regression in tracking stability since it set the tracking mode to free if there were any camera position callbacks while the animator wasn't animating. Instead, since there can be a delay between the last frame of an animation and the spurious callback it introduces sometimes, still update the position but don't set the tracking mode to free unless there was movement beyond a threshold as before.