[complass] Wrap camera controller in a state manager
The Android camera impl only allows one concurrent instance at a time; otherwise, disposing one controller disposes the shared camera that may be used by another.

The state management logic for the camera is complex, so pull it into its own utility.

There may still be rare race conditions where the camera gets stuck in the wrong orientation, possibly if the app lifecycle is paused and resumed while an initialization is taking place and orientation is changing, and the events line up in exactly the wrong way. This may be currently extremely difficult to test and it's not clear what the correct mitigation is; it may require sequentially waiting for the stale initialization to complete, disposing it, and then reinitializing, which seems unfortunate.

Fixes #20