[test] Test camera orientation reset on nogo select
When a nogo region is selected, if the camera is rotated, it rotates back to north. This has exposed a bug in the past where the camera control icon would stay stuck on the free-orientation compass rather than reset to the reticle.

This adds an integration test to cover this case.

This also fails tests on `waitFor` timeouts for transparency, and mocks out the WMS tile http client to produce noncompleting futures rather than null responses to quiet errors from native Maps.