[info] Keep info panel tabs NeverScrollable
There was an edge case where the entire info window could be scrolled, tabs and all, while collapsed using a mouse wheel.

Also takes the opportunity to use actual Material tabs, which are large enough to make it easier to expand the collapsed position at the bottom of the screen, particularly without counting as an app switch gesture.