[info] Don't flip station priorities for panels other than Tides and Details
Flipping marker z-order is an expensive operation that blocks the main thread, which would require a change in the maps plugin to break apart. For now, since the 'Details' and 'Plot' panels are adjacent and the 'Plot' panel shouldn't really care what the map marker z-order is, skip the flip for panels other than the 'Tides' and 'Details' panels.

Fixes #29