[plot][main] Fix map fit to track while panel expanded
Using a viewport to clip the map was offstaging the map while occluded, which causes the JS Maps API camera animate-to-bounds to zoom out extremely. Use an OverflowBox instead.

Breaks out integration tests meant to run on all platforms. Keeps the rotation tests as native tests to benefit from screenshots as rotation doesn't work on web in the current configuration.