[test] Attempt to deflake native test
The native test flaked waiting for the map orientation lock icon to change to a reticle after tapping on the compass arrow post-drag. This was probably because drag events route through the native side before being called back into Flutter, which may have missed a frame update on the slow emulator and ended up being skipped by pumpAndSettle.

Since there's not really a great way to sync with the native side, pump for one second.