[test] Attempt to deflake native test
Add a pumpUntilMapSettled call in the nogo bounds test in case the emulator has fallen behind updating for the drag.

Since pumpUntilMapSettled first waits for the camera position to change, it's not clear why it's not sometimes hanging, but if it starts doing so, we can tweak it so that a variant skips that wait, and perhaps waits for several frames without changes.