[server] Caching
Write locks are being held for the duration of a fetch, which can be long. Two possibilities for dealing with this include downloading to a temporary file or using a memory cache. This goes with a memory cache to also save on disk IO costs. For simplicity, there's still a cache miss case under heavy load where a write lock may block reads. Also for simplcity since latlngs are infrequently read and seldom written, they are not cached.