Initial Dart 3 bump
A pinned hash for my file.dart fork got garbage collected after I made changes in preparation for a pull request. Rather than pin to the current hash, I decided to use the official version since the PR's been merged.

Unfortunately this meant that it was Dart 3 time (at least partially), which brings fresh new deprecations that we'll want to deal with eventually.