[server] Update file.dart pin
The previously pinned hash was garbage collected due to updates to a pull request. Although this PR has been merged, upgrading it requires upgrading the Flutter and Dart SDKs, which needs to be done carefully. I attempted to
upgrade to master, but flutter_layout_grid via flutter_map is not compatible with flutter/flutter 5ab5d82.